+++ src/Confluent.SchemaRegistry.Serdes/AvroDeserializer.cs
@@ -36,7 +36,7 @@ namespace Confluent.SchemaRegistry.Serdes
     ///       bytes 1-4:        Unique global id of the Avro schema that was used for encoding (as registered in Confluent Schema Registry), big endian.
     ///       following bytes:  The serialized data.
     /// </remarks>
-    public class AvroDeserializer<T> : ITaskDeserializer<T>
+    public class AvroDeserializer<T> : IAsyncDeserializer<T>
     {
         private IAvroDeserializerImpl<T> deserializerImpl;
 
@@ -76,26 +76,28 @@ namespace Confluent.SchemaRegistry.Serdes
         ///     Deserialize an object of type <typeparamref name="T"/>
         ///     from a byte array.
         /// </summary>
-        /// <param name="topic">
-        ///     The topic associated with the message the raw data
-        ///     is associated with.
-        /// </param>
         /// <param name="data">
-        ///     The data to deserialize.
+        ///     The raw byte data to deserialize.
+        /// </param>
+        /// <param name="isNull">
+        ///     True if this is a null value.
+        /// </param>
+        /// <param name="messageAncillary">
+        ///     Properties of the message the data is associated with in
+        ///     addition to the key or value.
+        /// </param>
+        /// <param name="source">
+        ///     The TopicPartition from which the message was consumed.
         /// </param>
         /// <param name="isKey">
-        ///     True if deserializing message key data, false if
-        ///     deserializing message value data.
-        /// </param>
-        /// <param name="headers">
-        ///     The headers of the message associated with this
-        ///     value.
+        ///     True if deserializing the message key, false if deserializing the
+        ///     message value.
         /// </param>
         /// <returns>
         ///     A <see cref="System.Threading.Tasks.Task" /> that completes
         ///     with the deserialized value.
         /// </returns>
-        public async Task<T> Deserialize(byte[] data, bool isKey, string topic, Headers headers)
+        public async Task<T> DeserializeAsync(ReadOnlyMemory<byte> data, bool isNull, bool isKey, MessageAncillary messageAncillary, TopicPartition source)
         {
             try
             {
@@ -106,7 +108,8 @@ namespace Confluent.SchemaRegistry.Serdes
                         : new SpecificDeserializerImpl<T>(schemaRegistryClient);
                 }
 
-                return await deserializerImpl.Deserialize(topic, data.ToArray());
+                // TODO: change this interface such that it takes ReadOnlyMemory<byte>, not bytet[].
+                return await deserializerImpl.Deserialize(source.Topic, isNull ? null : data.ToArray());
             }
             catch (AggregateException e)
             {